+++ test/assertFileEnv/.test-dev.test.ts
@@ -1,0 +1,2 @@
+import { testRun } from './.testRun'
+testRun('npm run dev')
