+++ test/assertFileEnv/.testCiJob.json
@@ -1,0 +1,1 @@
+{ "name": "Examples React" }
