+++ test/assertFileEnv/pages/+config.h.js
@@ -1,0 +1,2 @@
+import vikeReact from 'vike-react'
+export default { extends: vikeReact }
