+++ test/assertFileEnv/pages/index/+Page.server.jsx
@@ -1,0 +1,13 @@
+export default Page
+
+import React from 'react'
+import { Counter } from './Counter.client'
+
+function Page() {
+  return (
+    <>
+      <h1>Test page</h1>
+      <Counter />
+    </>
+  )
+}
