+++ test/assertFileEnv/pages/index/Counter.client.jsx
@@ -1,0 +1,9 @@
+import React, { useState } from 'react'
+
+export { Counter }
+
+function Counter() {
+  const [count, setCount] = useState(0)
+
+  return <button onClick={() => setCount((count) => count + 1)}>Counter {count}</button>
+}
