+++ test/assertFileEnv/vite.config.js
@@ -1,0 +1,3 @@
+import react from '@vitejs/plugin-react'
+import vike from 'vike/plugin'
+export default { plugins: [react(), vike()] }
