--- test/misc/pages/forbidden-import/+config.h.ts
@@ -1,5 +1,0 @@
-import type { Config } from 'vike/types'
-
-export default {
-  Page: 'import:./Page.jsx'
-} satisfies Config
