--- test/misc/pages/forbidden-import/ClientOnly.client.jsx
@@ -1,7 +1,0 @@
-export default ClientOnly
-
-import React from 'react'
-
-function ClientOnly() {
-  return <div>Client Only Component</div>
-}
