--- test/misc/pages/forbidden-import/Page.jsx
@@ -1,11 +1,0 @@
-export default Page
-
-import React from 'react'
-
-function Page() {
-  return (
-    <div>
-      <h1>This is page for forbidden import.</h1>
-    </div>
-  )
-}
