--- test/misc/pages/index/Counter.jsx
@@ -1,12 +1,0 @@
-export { Counter }
-
-import React, { useState } from 'react'
-
-function Counter() {
-  const [count, setCount] = useState(0)
-  return (
-    <button type="button" onClick={() => setCount((count) => count + 1)}>
-      Counter {count}
-    </button>
-  )
-}
